* kpasswd.c (get_name_from_passwd_file): Do not cast return from 	getuid() to int as argument to getpwuid.
	* ksetpwd.c: Include time.h for time() prototype.
	(main): Return 0 on success.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@16417 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/clients/kpasswd/kpasswd.c
@@ -21,7 +21,7 @@ void get_name_from_passwd_file(program_name, kcontext, me)
 {
     struct passwd *pw;
     krb5_error_code code;
-    if ((pw = getpwuid((int) getuid()))) {
+    if ((pw = getpwuid(getuid()))) {
 	if ((code = krb5_parse_name(kcontext, pw->pw_name, me))) {
 	    com_err (program_name, code, "when parsing name %s", pw->pw_name);
 	    exit(1);
